Validate delete endpoint

+++ app/mgmt/[secret]/api/route.ts
@@ -1,16 +1,20 @@
 import { PARTYKIT_URL } from "@/app/env";
 import { NextRequest } from "next/server";
 
+const PARTY_SECRET = process.env.PARTY_SECRET;
+
 export const DELETE = (req: NextRequest) => {
-  const id = req.nextUrl.searchParams.get("id");
-  if (id) {
-    return fetch(`${PARTYKIT_URL}/party/${id}`, {
-      method: "DELETE",
-      headers: {
-        Authorization: process.env.PARTY_SECRET!,
-      },
-    });
+  if (PARTY_SECRET && req.nextUrl.pathname.includes(PARTY_SECRET)) {
+    const id = req.nextUrl.searchParams.get("id");
+    if (id) {
+      return fetch(`${PARTYKIT_URL}/party/${id}`, {
+        method: "DELETE",
+        headers: {
+          Authorization: PARTY_SECRET,
+        },
+      });
+    }
   }
 
-  return new Response("Missing id", { status: 400 });
+  return new Response("Bad request", { status: 400 });
 };
